Organize scripts folder structure
