+++ hardhat.config.ts
@@ -6,7 +6,7 @@ import '@nomicfoundation/hardhat-toolbox'
 import 'hardhat-contract-sizer'
 import './scripts/deploy'
 import './scripts/wallet'
-import './scripts/mintPlatformIdForAddress'
+import './scripts/tasks/protocol/mintPlatformIdForAddress'
 import './scripts/tasks/protocol/addArbitrator'
 import './scripts/tasks/protocol/removeArbitrator'
 import { Network } from './scripts/config'
