+++ scripts/playground/disputeResolution/3-submit-evidence.ts
@@ -1,7 +1,7 @@
 import { ethers } from 'hardhat'
 import { ConfigProperty, get } from '../../../configManager'
 import { Network } from '../../config'
-import postToIPFS from '../../ipfs'
+import postToIPFS from '../../utils/ipfs'
 import { transactionId } from './constants'
 
 const hre = require('hardhat')
